REF: Rename logging_handlers module to log_handlers
